refactor: migrate analytics models and stabilize database registry
