+++ api/config.py
@@ -1,8 +1,6 @@
-"""
-API Configuration
-"""
-from functools import lru_cache
+import os
 import tempfile
+from functools import lru_cache
 from pathlib import Path
 from typing import Optional
 from pydantic_settings import BaseSettings
